refactor(edition): extract EditionHeader component
Move TopBar + FestivalIndicator + IdentityRow scroll wiring out of
EditionLayout into a dedicated EditionHeader component.
